fixed curses build

+++ src/hal/curses/simple_console_backing.rs
@@ -43,7 +43,7 @@ impl SimpleConsoleBackend {
                 let cp_fg = find_nearest_color(t.fg, &platform.platform.color_map);
                 let cp_bg = find_nearest_color(t.bg, &platform.platform.color_map);
                 let pair = (cp_bg * 16) + cp_fg;
-                window.attrset(pancurses::COLOR_PAIR(pair as u64));
+                window.attrset(pancurses::COLOR_PAIR(pair as u32));
                 window.mvaddch(
                     height as i32 - (y as i32 + 1),
                     x as i32,
